move loading into separate plugins
