+++ plugins/loading/README.md
@@ -1,0 +1,3 @@
+# Rematch-Loading
+
+WIP
