+++ plugins/loading/src/index.js
